releases/current/orchestrator/requirements.txt
paho-mqtt==2.1.0
psutil==5.9.8
PyYAML==6.0.2
jsonschema==4.22.0
rich==13.7.1
requests==2.32.3
tzdata==2024.1
